+++ projects/XDecoder/configs/xdecoder-tiny_zeroshot_open-vocab-instance.py
@@ -1,13 +1,7 @@
 _base_ = 'xdecoder-tiny_zeroshot_open-vocab-semseg.py'
 
-model = dict(task='instance',
-             test_cfg=dict(
-                 nms_pre=1000,
-                 min_bbox_size=0,
-                 score_thr=0.05,
-                 nms=dict(type='nms', iou_threshold=0.5),
-                 max_per_img=100)
-             )
+model = dict(head=dict(task='instance'),
+             test_cfg=dict(max_per_img=100))
 
 test_pipeline = [
     dict(type='LoadImageFromFile', imdecode_backend='pillow', backend_args=_base_.backend_args),
