+++ projects/XDecoder/configs/xdecoder-tiny_zeroshot_ref-caption.py
@@ -1,11 +1,11 @@
 _base_ = 'xdecoder-tiny_zeroshot_open-vocab-semseg.py'
 
-model = dict(task='ref-captioning')
+model = dict(head=dict(task='ref-caption'))
 
-grounding_resize_scale = 512
+grounding_scale = 512
 
 test_pipeline = [
-    dict(type='LoadImageFromFile',imdecode_backend='pillow', backend_args=_base_.backend_args),
+    dict(type='LoadImageFromFile', imdecode_backend='pillow', backend_args=_base_.backend_args),
     dict(type='FixScaleResize',
          scale=224,
          keep_ratio=True,
