+++ projects/XDecoder/configs/xdecoder-tiny_zeroshot_ref-semseg.py
@@ -1,3 +1,3 @@
 _base_ = 'xdecoder-tiny_zeroshot_open-vocab-semseg.py'
 
-model = dict(task='ref-semseg')
+model = dict(head=dict(task='ref-semseg'))
